move validation into legacy adapter
